made project structure
