+++ Smart Alarm.xcodeproj/xcuserdata/andreynnt.xcuserdatad/xcschemes/xcschememanagement.plist
