+++ Smart Alarm.xcodeproj/xcuserdata/evgenij.xcuserdatad/xcschemes/design.xcscheme
