+++ Smart Alarm.xcodeproj/xcuserdata/evgenij.xcuserdatad/xcschemes/xcschememanagement.plist
