+++ Smart Alarm/Alarms View Controller/AlarmCell.swift
@@ -8,16 +8,13 @@
 import UIKit
 
 
-class Cell: UITableViewCell {
-    
-    
+class AlarmCell: UITableViewCell {
     @IBOutlet weak var arrivetime: UILabel!
     @IBOutlet weak var arriveplace: UILabel!
     @IBOutlet weak var timeonfees: UILabel!
-    
     @IBOutlet weak var arriveTimeMin: UILabel!
+    
     override func awakeFromNib() {
         super.awakeFromNib()
     }
-    
 }
