+++ Smart Alarm/Assets.xcassets/AppIcon.appiconset/Contents.json
