+++ Smart Alarm/Assets.xcassets/Contents.json
