+++ Smart Alarm/Assets.xcassets/images.imageset/Contents.json
