+++ Smart Alarm/Info.plist
