+++ Smart Alarm/MapViewController.swift
@@ -1,0 +1,19 @@
+//
+//  MapViewController.swift
+//  Smart Alarm
+//
+//  Created by Андрей Бабков on 06/12/2018.
+//  Copyright © 2018 Tonya. All rights reserved.
+//
+
+import UIKit
+
+
+// Вьюха, в которой будет карта
+class MapViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+    
+}
