+++ Smart Alarm/PrepareSettingsViewController.swift
@@ -1,0 +1,18 @@
+//
+//  PrepareSettingsViewController.swift
+//  Smart Alarm
+//
+//  Created by Андрей Бабков on 06/12/2018.
+//  Copyright © 2018 Tonya. All rights reserved.
+//
+
+import UIKit
+
+
+// Вьюха, в которой будем устанавливать время на сборы
+class PrepareSettingsViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
