+++ Smart Alarm/TypeOfTansportViewController.swift
@@ -1,0 +1,19 @@
+//
+//  TypeOfTansportViewController.swift
+//  Smart Alarm
+//
+//  Created by Андрей Бабков on 06/12/2018.
+//  Copyright © 2018 Tonya. All rights reserved.
+//
+
+import UIKit
+
+
+// Вьюха, в которой выбираем тип транспорта, на котором отправимся в путь
+class TypeOfTansportViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+
+}
